Handles resources without a source file.
Some extensions (such as middleman-automatic-clowncar) create sitemap resources that do not have a corresponding source_file. This was causing the "p.source_file.include?(name)" expression to raise an error (as the nil object does not have an include? method).

This small change will only evaluate the expression when the source_file is not nil.